tar: rename files and reorganize functions...
...to better represent what's required of each major implementation
